Сделал по PEP, разделил логику и main
